# Conflicts:
#	features/merchant/official_store/src/main/java/com/tokopedia/officialstore/official/presentation/OfficialHomeFragment.kt
#	features/merchant/official_store/src/main/java/com/tokopedia/officialstore/official/presentation/viewmodel/OfficialStoreHomeViewModel.kt

+++ features/merchant/official_store/src/main/java/com/tokopedia/officialstore/official/presentation/OfficialHomeFragment.kt
@@ -61,6 +61,7 @@ import com.tokopedia.recommendation_widget_common.listener.RecommendationListene
 import com.tokopedia.recommendation_widget_common.presentation.model.RecommendationItem
 import com.tokopedia.remoteconfig.FirebaseRemoteConfigImpl
 import com.tokopedia.remoteconfig.RemoteConfig
+import com.tokopedia.remoteconfig.RemoteConfigKey
 import com.tokopedia.unifycomponents.Toaster
 import com.tokopedia.usecase.coroutines.Fail
 import com.tokopedia.usecase.coroutines.Success
@@ -212,7 +213,7 @@ class OfficialHomeFragment :
         observeProductRecommendation()
         observeFeaturedShopSuccessDC()
         observeFeaturedShopRemoveDC()
-        observeRecomwidget()
+        observeRecomWidget()
         initLocalChooseAddressData()
         resetData()
         loadData()
@@ -223,12 +224,16 @@ class OfficialHomeFragment :
 
     }
 
-    private fun observeRecomwidget() {
-        viewModel.recomWidget.observe(viewLifecycleOwner, {
+    private fun observeRecomWidget() {
+        viewModel.recomWidget.observe(viewLifecycleOwner) {
             when (it) {
                 is Success -> {
                     swipeRefreshLayout?.isRefreshing = false
-                    officialHomeMapper.mappingRecomWidget(it.data, adapter)
+                    if (!isEligibleForDisableBestSellerWidget()) {
+                        officialHomeMapper.mappingRecomWidget(it.data) { newDataList ->
+                            adapter?.submitList(newDataList)
+                        }
+                    }
                 }
                 is Fail -> {
                     swipeRefreshLayout?.isRefreshing = false
@@ -236,7 +241,7 @@ class OfficialHomeFragment :
                 }
 
             }
-        })
+        }
     }
 
     override fun onPause() {
@@ -803,7 +808,7 @@ class OfficialHomeFragment :
     }
 
     private fun observeBannerData() {
-        viewModel.officialStoreBannersResult.observe(viewLifecycleOwner, {
+        viewModel.officialStoreBannersResult.observe(viewLifecycleOwner) {
             val resultValue = it.second
 
             val shouldShowErrorMessage = it.first
@@ -813,7 +818,12 @@ class OfficialHomeFragment :
                         this.currentBannerData = resultValue.data
                         removeLoading(resultValue.data.isCache)
                         swipeRefreshLayout?.isRefreshing = false
-                        officialHomeMapper.mappingBanners(resultValue.data, adapter, category?.title)
+                        officialHomeMapper.mappingBanners(
+                            resultValue.data,
+                            adapter,
+                            category?.title,
+                            isEligibleForDisableMappingBanner()
+                        )
                     }
                 }
                 is Fail -> {
@@ -824,7 +834,7 @@ class OfficialHomeFragment :
                 }
             }
             bannerPerformanceMonitoring.stopTrace()
-        })
+        }
     }
 
     private fun observeBenefit() {
@@ -902,24 +912,31 @@ class OfficialHomeFragment :
     }
 
     private fun observeFeaturedShopSuccessDC() {
-        viewModel.featuredShopResult.observe(viewLifecycleOwner, {
-            when(it) {
+        viewModel.featuredShopResult.observe(viewLifecycleOwner) {
+            when (it) {
                 is Success -> {
                     //update UI
-                    officialHomeMapper.updateFeaturedShopDC(it.data) { newDataList ->
-                        adapter?.submitList(newDataList)
+                    if (!isEligibleForDisableShopWidget()) {
+                        officialHomeMapper.updateFeaturedShopDC(
+                            it.data
+                        ) { newDataList ->
+                            adapter?.submitList(newDataList)
+                        }
                     }
                 }
+                else -> {}
             }
-        })
+        }
     }
 
     private fun observeFeaturedShopRemoveDC() {
-        viewModel.featuredShopRemove.observe(viewLifecycleOwner, {
-            officialHomeMapper.removeFeaturedShopDC(it) { newDataList ->
-                adapter?.submitList(newDataList)
+        viewModel.featuredShopRemove.observe(viewLifecycleOwner) {
+            if (!isEligibleForDisableRemoveShopWidget()) {
+                officialHomeMapper.removeFeaturedShopDC(it) { newDataList ->
+                    adapter?.submitList(newDataList)
+                }
             }
-        })
+        }
     }
 
     private fun showErrorNetwork(t: Throwable) {
@@ -930,12 +947,20 @@ class OfficialHomeFragment :
         }
     }
 
+    private fun removeRecomWidget() {
+        if (!isEligibleForDisableRemoveBestSellerWidget()) {
+            officialHomeMapper.removeRecomWidget {
+                adapter?.submitList(it)
+            }
+        }
+    }
+
     private fun setListener() {
         setLoadMoreListener()
         swipeRefreshLayout?.setOnRefreshListener {
             counterTitleShouldBeRendered = 0
             officialHomeMapper.removeRecommendation(adapter)
-            officialHomeMapper.removeRecomWidget(adapter)
+            officialHomeMapper.removeRecomWidget()
             officialHomeMapper.removeTopAdsHeadlineWidget(adapter)
             loadData(true)
             viewModel.resetShopWidgetImpressionCount()
@@ -1006,7 +1031,7 @@ class OfficialHomeFragment :
     private fun showSuccessAddWishlist() {
         activity?.let { activity ->
             val view = activity.findViewById<View>(android.R.id.content) ?: return
-            val message = getString(R.string.msg_success_add_wishlist)
+            val message = getString(com.tokopedia.officialstore.R.string.msg_success_add_wishlist)
 
             Snackbar.make(view, message, Snackbar.LENGTH_LONG)
                     .setAction("Lihat Wishlist") { RouteManager.route(activity, ApplinkConst.WISHLIST) }
@@ -1018,7 +1043,7 @@ class OfficialHomeFragment :
     private fun showSuccessRemoveWishlist() {
         activity?.let {
             val view = it.findViewById<View>(android.R.id.content) ?: return
-            val message = getString(R.string.msg_success_remove_wishlist)
+            val message = getString(com.tokopedia.officialstore.R.string.msg_success_remove_wishlist)
 
             Snackbar.make(view, message, Snackbar.LENGTH_LONG).show()
         }
@@ -1093,6 +1118,51 @@ class OfficialHomeFragment :
         }
         return false
 
+    }
+
+    private fun isEligibleForDisableShopWidget(): Boolean {
+        return try {
+            return remoteConfig?.getBoolean(RemoteConfigKey.DISABLE_OFFICIAL_STORE_SHOP_WIDGET)
+                ?: false
+        } catch (e: Exception) {
+            false
+        }
+    }
+
+    private fun isEligibleForDisableBestSellerWidget(): Boolean {
+        return try {
+            return remoteConfig?.getBoolean(RemoteConfigKey.DISABLE_OFFICIAL_STORE_BEST_SELLER_WIDGET)
+                ?: false
+        } catch (e: Exception) {
+            false
+        }
+    }
+
+    private fun isEligibleForDisableMappingBanner(): Boolean {
+        return try {
+            return remoteConfig?.getBoolean(RemoteConfigKey.DISABLE_OFFICIAL_STORE_MAPPING_BANNERS)
+                ?: false
+        } catch (e: Exception) {
+            false
+        }
+    }
+
+    private fun isEligibleForDisableRemoveBestSellerWidget(): Boolean {
+        return try {
+            return remoteConfig?.getBoolean(RemoteConfigKey.DISABLE_OFFICIAL_STORE_REMOVE_BEST_SELLER_WIDGET)
+                ?: false
+        } catch (e: Exception) {
+            false
+        }
+    }
+
+    private fun isEligibleForDisableRemoveShopWidget(): Boolean {
+        return try {
+            return remoteConfig?.getBoolean(RemoteConfigKey.DISABLE_OFFICIAL_STORE_REMOVE_SHOP_WIDGET)
+                ?: false
+        } catch (e: Exception) {
+            false
+        }
     }
 
     private fun RecommendationItem.createProductCardOptionsModel(position: Int): ProductCardOptionsModel {
